// src/lib/types/collections.ts
export interface UserDataDocumentsType extends Models.Row {
  /**
   * The avatar ID of the user.
   */
  avatarId: string | null
  /**
   * The banner ID of the user.
   */
  profileBannerId: string | null
  /**
   * The user's status.
   */
  status: string | null
  /**
   * The user's display name.
   */
  displayName: string
  /**
   * The user's bio.
   */
  bio: string | null
  /**
   * The user's birthday.
   * @example '2000-01-01T00:00:00:000Z'
   */
  birthday: string | null
  /**
   * The user's profile URL.
   */
  profileUrl: string
  /**
   * The user's pronouns.
   */
  pronouns: string | null
  /**
   * The user's discord username. (Not the ID)
   */
  discordname: string | null
  /**
   * The user's telegram name.
   */
  telegramname: string | null
  /**
   * The user's furaffinity name.
   */
  furaffinityname: string | null
  /**
   * The user's x/twitter name.
   */
  X_name: string | null
  /**
   * The user's twitch name.
   */
  twitchname: string | null
  /**
   * The user's bluesky name.
   */
  blueskyname: string | null
  /**
   * The user's location.
   */
  location: string | null
  /**
   * The user's badges.
   */
  badges: string[]
  /**
   * The user's followers count.
   */
  followersCount: number
  /**
   * The user's following count.
   */
  followingCount: number
}
// ...
export interface AnnouncementDataType {
  total: number
  rows: AnnouncementDocumentsType[]
}

/**
 * This data is returned from the API within the `documents` array.
 * @see AnnouncementDataType
 * @see UserDataType
 * @interface
 * @since 2.0.0
 */
export interface AnnouncementDocumentsType extends Models.Row {
  /**
   * The title of the announcement.
   */
  title: string | null
  /**
   * The side text of the announcement.
   */
  sideText: string | null
  /**
   * The description of the announcement.
   */
  description: string
  /**
   * Until when the announcement is valid.
   */
  validUntil: Date
  /**
   * The user data of the user that created the announcement.
   */
  userData: UserDataDocumentsType
}